requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
pydantic>=1.10.7
numpy>=1.24.2
pandas>=2.0.0
scikit-learn>=1.2.2
python-dotenv>=1.0.0
httpx>=0.24.0
pytest>=7.3.1
